[NETBEANS-4294] . Refactored plugin to remove most of the listeners, wrong WeakReferecnes, opted to use Lifecicle saveall option and refactored the Options dialog as review discussion; . Removed all OmegaT related files; . Removed help files and translations;
