--- ide/editor.autosave/build-i18n.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
-
--->
-<project name="org.netbeans.modules.editor.autosave" default="netbeans" basedir=".">
-    <description>Builds, tests, and runs the project org.netbeans.modules.editor.autosave.</description>
-    <import file="nbproject/build-impl.xml"/>
-    <property file="i18n.properties"/>
-    
-    <target name="-i18n-init">
-        <mkdir dir="${translatedfiles.basedir}/src"/>
-        <mkdir dir="${translatedfiles.basedir}/javahelp"/>
-    </target>
-    
-    <target name="-pre-javahelp" depends="-i18n-init">
-        <copy todir="${basedir}/build/javahelp" preservelastmodified="true" verbose="true">
-            <fileset dir="${translatedfiles.basedir}/javahelp">
-                <include name="**/*.*"/>
-            </fileset>
-        </copy>
-    </target>
-    
-    <target name="javahelp" depends="-pre-javahelp,harness.javahelp"/>
-    
-    <target name="-pre-jar-prep" depends="-i18n-init">
-        <copy todir="${build.classes.dir}" preservelastmodified="true" verbose="true">
-            <fileset dir="${translatedfiles.basedir}/src">
-                <include name="**/*.*"/>
-            </fileset>
-        </copy>
-    </target>
-    
-    <target name="jar-prep" depends="-pre-jar-prep,projectized-common.jar-prep"/>
-    
-    <target name="update-omegat-source" description="Update i18n OmegaT source">
-        <copy todir="${i18n.basedir}/${i18n.default.name}/source" preservelastmodified="false" verbose="true" overwrite="true">
-            <fileset dir=".">
-                <include name="src/**/*.properties"/>
-                <include name="src/**/*.htm*"/>
-                
-                <include name="javahelp/**/*.properties"/>
-                <include name="javahelp/**/*.htm*"/>
-                <include name="javahelp/**/*toc.xml"/>
-            </fileset>
-        </copy>
-    </target>
-</project>
