--- ide/editor.autosave/i18n.properties
@@ -1,23 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-#Basedir for OmegaT projects.
-i18n.basedir=${basedir}/i18n
-#Default OmegaT project name. This folder will have its sources updated.
-i18n.default.name=omegat-autosave
-#This folder will have all translated files.
-translatedfiles.basedir=translatedfiles
