--- ide/editor.autosave/i18n/omegat-ja/omegat-ja-level2.tmx
@@ -1,141 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
-
--->
-<!DOCTYPE tmx SYSTEM "tmx14.dtd">
-<tmx version="1.4">
-  <header
-    creationtool="OmegaT"
-    creationtoolversion="1.6.1"
-    segtype="paragraph"
-    o-tmf="OmegaT TMX"
-    adminlang="EN-US"
-    srclang="EN-US"
-    datatype="plaintext"
-  >
-  </header>
-  <body>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>About Save Automatically</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>自動保存について</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>nbdocs:/org/netbeans/modules/usersguide/ide.css</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>nbdocs:/org/netbeans/modules/usersguide/ja/ide.css</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>This module provides the functionality to automatically save all yours unsaved opened files.</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>このモジュールは編集中のすべてのファイルを保存する機能を提供します。</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>It is possible too define if this feature will be enabled or not, and the interval between save actions, by the appropriate option in <bpt i='0' x='0'>&lt;b0&gt;</bpt>Tools &gt; Options &gt; Miscellaneous &gt; Save Automatically<ept i='0'>&lt;/b0&gt;</ept>. To enable/disable this feature, is possible use the <bpt i='1' x='1'>&lt;b1&gt;</bpt>File &gt; <ph x='2'>&lt;i2/&gt;</ph> Save Automatically<ept i='1'>&lt;/b1&gt;</ept> option too.</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>この機能の有効/無効、自動保存の間隔はオプションダイアログで設定することができます。オプションは <bpt i='0' x='0'>&lt;b0&gt;</bpt>ツール &gt; オプション &gt; その他 &gt; 自動保存<ept i='0'>&lt;/b0&gt;</ept>で設定できます。有効/無効を設定するには <bpt i='1' x='1'>&lt;b1&gt;</bpt>ファイル &gt; <ph x='2'>&lt;i2/&gt;</ph>自動保存<ept i='1'>&lt;/b1&gt;</ept> オプションも使用できます。</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>Save Automatically</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>自動保存</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>Settings for save automatically feature</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>自動保存の設定</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>Editing</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>編集</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>Save automatically unsaved files between a predefined interval.
-The module options can be changed on Tools &gt; Options &gt; Miscellaneous &gt; Save Automatically section. It is possible too activate/deactivate this feature by the File &gt; Save Automatically option.</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>未保存のファイルを設定した間隔で自動的に保存します。モジュールのオプションは ツール &gt; オプション &gt; その他 &gt; 自動保存 で設定できます。有効/無効を設定するには ファイル &gt; 自動保存 オプションも使用できます。</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>Save files every</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>保存する間隔: </seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>minute(s).</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>分毎</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>Define the interval between save operations.</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>保存する間隔を設定します。</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>Use save automatically feature.</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>自動保存を有効</seg>
-      </tuv>
-    </tu>
-    <tu>
-      <tuv xml:lang="EN-US">
-        <seg>Save automatically unsaved files between a predefined interval.</seg>
-      </tuv>
-      <tuv xml:lang="JA">
-        <seg>未保存のファイルを自動保存する間隔。</seg>
-      </tuv>
-    </tu>
-  </body>
-</tmx>
