--- ide/editor.autosave/i18n/omegat-pt_BR/omegat/project_stats.txt
@@ -1,41 +1,0 @@
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-
-Estatística do projeto
-
-Total:
-	Segmentos	25
-	Palavras	185
-	Caracteres (sem espaços)	1226
-	Caracteres (incluindo espaços)	1368
-Únicos:
-	Segmentos	21
-	Palavras	168
-	Caracteres (sem espaços)	1106
-	Caracteres (incluindo espaços)	1235
-Segmentos únicos restantes:
-	Segmentos	4
-	Palavras	38
-	Caracteres (sem espaços)	257
-	Caracteres (incluindo espaços)	270
-
-Estatística do arquivo individual:
-
-Nome do arquivo	Total de palavras	Palavras restantes para traduzir	Total de caracteres (sem espaços)	Caracteres restantes (sem espaços)	Total de caracteres (incluindo espaços)	Caracteres restantes (incluindo espaços)
-javahelp\org\netbeans\modules\autosave\docs\autosave-about.html	69	7	433	60	494	60
-javahelp\org\netbeans\modules\autosave\docs\autosave-toc.xml	42	31	310	197	329	210
-src\org\netbeans\modules\autosave\Bundle.properties	74	0	483	0	545	0
