--- ide/editor.autosave/i18n/omegat-pt_BR/source/src/org/netbeans/modules/autosave/Bundle.properties
@@ -1,34 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-AdvancedOption_DisplayName=Save Automatically
-AdvancedOption_Tooltip=Settings for save automatically feature
-CTL_AutoSave=Save Automatically
-OpenIDE-Module-Display-Category=Editing
-OpenIDE-Module-Long-Description=\
-    Save automatically unsaved files between a predefined interval.\n\
-    The module options can be changed on Tools > Options > Miscellaneous > Save Automatically section. \
-    It is possible too activate/deactivate this feature by the File > Save Automatically option.
-OpenIDE-Module-Name=AutoSave Module
-
-AutoSavePanel.jLabel1.text=Save files every
-AutoSavePanel.jLabel2.text=minute(s).
-AutoSavePanel.spnMinutes.toolTipText=Define the interval between save operations.
-
-AutoSavePanel.chkUseFeature.actionCommand=Use save automatically feature.
-AutoSavePanel.chkUseFeature.text=Use save automatically feature.
-OpenIDE-Module-Short-Description=Save automatically unsaved files between a predefined interval.
