--- ide/editor.autosave/javahelp/org/netbeans/modules/autosave/docs/autosave-hs.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
-
--->
-<!DOCTYPE helpset PUBLIC "-//Sun Microsystems Inc.//DTD JavaHelp HelpSet Version 2.0//EN" "http://java.sun.com/products/javahelp/helpset_2_0.dtd">
-<helpset version="2.0">
-    <title>Save Automatically Help</title>
-    <maps>
-        <homeID>org.netbeans.modules.editor.autosave.about</homeID>
-        <mapref location="autosave-map.xml"/>
-    </maps>
-    <view mergetype="javax.help.AppendMerge">
-        <name>TOC</name>
-        <label>Table of Contents</label>
-        <type>javax.help.TOCView</type>
-        <data>autosave-toc.xml</data>
-    </view>
-    <view mergetype="javax.help.AppendMerge">
-        <name>Index</name>
-        <label>Index</label>
-        <type>javax.help.IndexView</type>
-        <data>autosave-idx.xml</data>
-    </view>
-    <view>
-        <name>Search</name>
-        <label>Search</label>
-        <type>javax.help.SearchView</type>
-        <data engine="com.sun.java.help.search.DefaultSearchEngine">JavaHelpSearch</data>
-    </view>
-</helpset>
