+++ ide/editor.autosave/nbproject/project.properties
@@ -17,5 +17,5 @@
 
 javac.compilerargs=-Xlint -Xlint:-serial
 javac.source=1.8
-javadoc.title=Editor Save Automatically
+javadoc.title=Editor Autosave
 javahelp.hs=autosave-hs.xml
