+++ ide/editor.autosave/src/org/netbeans/modules/editor/autosave/AutoSavePanel.java
@@ -20,10 +20,6 @@ package org.netbeans.modules.editor.autosave;
 
 import java.awt.event.ItemEvent;
 import java.awt.event.ItemListener;
-import java.lang.ref.WeakReference;
-import java.util.prefs.PreferenceChangeEvent;
-import java.util.prefs.PreferenceChangeListener;
-import javax.swing.BorderFactory;
 import javax.swing.GroupLayout;
 import javax.swing.GroupLayout.Alignment;
 import javax.swing.JCheckBox;
@@ -31,7 +27,6 @@ import javax.swing.JLabel;
 import javax.swing.JSpinner;
 import javax.swing.LayoutStyle.ComponentPlacement;
 import javax.swing.SpinnerNumberModel;
-import org.netbeans.modules.editor.autosave.command.AutoSaveController;
 import org.openide.awt.Mnemonics;
 import org.openide.util.NbBundle;
 
@@ -41,30 +36,23 @@ final class AutoSavePanel extends javax.swing.JPanel {
         spnModel = new SpinnerNumberModel(10, 0, 999, 1);
 
         initComponents();
-
-        AutoSaveController.prefs().
-                addPreferenceChangeListener(new WeakReference<PreferenceChangeListener>(evt -> {
-                    controller.update();
-                    controller.changed();
-                }).get());
     }
 
     /**
      * This method is called from within the constructor to initialize the form. WARNING: Do NOT modify this code. The
      * content of this method is always regenerated by the Form Editor.
      */
+
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
         chkUseFeature = new JCheckBox();
         chkSaveOnFocusLost = new JCheckBox();
-        jLabel1 = new JLabel();
         spnMinutes = new JSpinner();
         jLabel2 = new JLabel();
 
-        Mnemonics.setLocalizedText(chkUseFeature, NbBundle.getMessage(AutoSavePanel.class, "AutoSavePanel.chkUseFeature.text")); // NOI18N
+        Mnemonics.setLocalizedText(chkUseFeature, NbBundle.getMessage(AutoSavePanel.class, "AutoSavePanel.jLabel1.text")); // NOI18N
         chkUseFeature.setActionCommand(NbBundle.getMessage(AutoSavePanel.class, "AutoSavePanel.chkUseFeature.actionCommand")); // NOI18N
-        chkUseFeature.setBorder(BorderFactory.createEmptyBorder(0, 0, 0, 0));
         chkUseFeature.addItemListener(new ItemListener() {
             public void itemStateChanged(ItemEvent evt) {
                 chkUseFeatureItemStateChanged(evt);
@@ -72,8 +60,6 @@ final class AutoSavePanel extends javax.swing.JPanel {
         });
 
         Mnemonics.setLocalizedText(chkSaveOnFocusLost, NbBundle.getMessage(AutoSavePanel.class,"AutoSavePanel.chkSaveOnFocusLost.text")); // NOI18N
-
-        Mnemonics.setLocalizedText(jLabel1, NbBundle.getMessage(AutoSavePanel.class, "AutoSavePanel.jLabel1.text")); // NOI18N
 
         spnMinutes.setModel(this.spnModel);
         spnMinutes.setToolTipText(NbBundle.getMessage(AutoSavePanel.class, "AutoSavePanel.spnMinutes.toolTipText")); // NOI18N
@@ -86,36 +72,29 @@ final class AutoSavePanel extends javax.swing.JPanel {
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap()
                 .addGroup(layout.createParallelGroup(Alignment.LEADING)
+                    .addComponent(chkSaveOnFocusLost)
                     .addGroup(layout.createSequentialGroup()
-                        .addGap(19, 19, 19)
-                        .addGroup(layout.createParallelGroup(Alignment.LEADING)
-                            .addGroup(layout.createSequentialGroup()
-                                .addComponent(jLabel1)
-                                .addPreferredGap(ComponentPlacement.RELATED)
-                                .addComponent(spnMinutes, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(ComponentPlacement.RELATED)
-                                .addComponent(jLabel2))
-                            .addComponent(chkSaveOnFocusLost)))
-                    .addComponent(chkUseFeature))
+                        .addComponent(chkUseFeature)
+                        .addPreferredGap(ComponentPlacement.RELATED)
+                        .addComponent(spnMinutes, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(ComponentPlacement.RELATED)
+                        .addComponent(jLabel2)))
                 .addContainerGap(GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(layout.createParallelGroup(Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap()
-                .addComponent(chkUseFeature)
-                .addPreferredGap(ComponentPlacement.RELATED)
-                .addComponent(chkSaveOnFocusLost)
-                .addPreferredGap(ComponentPlacement.RELATED)
                 .addGroup(layout.createParallelGroup(Alignment.BASELINE)
-                    .addComponent(jLabel1)
+                    .addComponent(chkUseFeature)
                     .addComponent(spnMinutes, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
                     .addComponent(jLabel2))
+                .addPreferredGap(ComponentPlacement.RELATED)
+                .addComponent(chkSaveOnFocusLost)
                 .addContainerGap(GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
     }// </editor-fold>//GEN-END:initComponents
 
     private void chkUseFeatureItemStateChanged(java.awt.event.ItemEvent evt) {//GEN-FIRST:event_chkUseFeatureItemStateChanged
-        this.chkSaveOnFocusLost.setEnabled(this.chkUseFeature.isSelected());
         this.spnMinutes.setEnabled(this.chkUseFeature.isSelected());
     }//GEN-LAST:event_chkUseFeatureItemStateChanged
 
@@ -134,13 +113,9 @@ final class AutoSavePanel extends javax.swing.JPanel {
         AutoSaveController.prefs().putInt(AutoSaveController.KEY_INTERVAL, spnModel.getNumber().intValue());
     }
 
-    boolean valid() {
-        return true;
-    }
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private JCheckBox chkSaveOnFocusLost;
     private JCheckBox chkUseFeature;
-    private JLabel jLabel1;
     private JLabel jLabel2;
     private JSpinner spnMinutes;
     // End of variables declaration//GEN-END:variables
