+++ ide/editor.autosave/src/org/netbeans/modules/editor/autosave/Bundle.properties
@@ -15,22 +15,20 @@
 # specific language governing permissions and limitations
 # under the License.
 
-OpenIDE-Module-Name=Editor Save Automatically
+OpenIDE-Module-Name=Editor Autosave
 
-AdvancedOption_DisplayName=Save Automatically
-CTL_AutoSave=Save Automatically
 OpenIDE-Module-Display-Category=Editing
 OpenIDE-Module-Long-Description=\
-    Save automatically unsaved files between a predefined interval.\n\
-    The module options can be changed on Tools > Options > Editor > Save Automatically section. \
-    It is possible too activate/deactivate this feature by the File > Save Automatically option.
-OpenIDE-Module-Name=Editor Save Automatically
+    Autosave unsaved files between a predefined interval.\n\
+    The module options can be changed on Tools > Options > Editor > Autosave section. \
+    It is possible too activate/deactivate this feature by the File > Autosave option.
+OpenIDE-Module-Name=Editor Autosave
 
 AutoSavePanel.jLabel1.text=Save files every
 AutoSavePanel.jLabel2.text=minute(s)
 AutoSavePanel.spnMinutes.toolTipText=Define the time interval between save operations.
 
-AutoSavePanel.chkUseFeature.actionCommand=Use save automatically feature.
-AutoSavePanel.chkUseFeature.text=Use save automatically feature
-OpenIDE-Module-Short-Description=Save automatically unsaved files between a predefined time interval.
+AutoSavePanel.chkUseFeature.actionCommand=Use autosave feature.
+AutoSavePanel.chkUseFeature.text=Use autosave feature
+OpenIDE-Module-Short-Description=Autosave unsaved files between a predefined time interval.
 AutoSavePanel.chkSaveOnFocusLost.text=Save files when focus is lost
