--- ide/editor.autosave/translatedfiles/javahelp/org/netbeans/modules/editor/autosave/docs/autosave-about_ja.html
@@ -1,44 +1,0 @@
-<!--
-
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
-
--->
-<html>
-    <head>
-        <title>自動保存について</title>
-        <link rel="stylesheet" href="nbdocs:/org/netbeans/modules/usersguide/ja/ide.css" type="text/css">
-        <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-    </head>
-    <body>
-        <h2>自動保存について</h2>
-         <p>このモジュールは編集中のすべてのファイルを保存する機能を提供します。</p>
-        <p>この機能の有効/無効、自動保存の間隔はオプションダイアログで設定することができます。オプションは <b>ツール &gt; オプション &gt; その他 &gt; 自動保存</b>で設定できます。有効/無効を設定するには <b>ファイル &gt; <img src="auto_save.png"/>自動保存</b> オプションも使用できます。</p>
-    </body>
-</html><!--
-Tip: to create a link which will open in an external web browser, try:
-<object classid="java:org.netbeans.modules.javahelp.BrowserDisplayer">
-    <param name="content" value="http://www.netbeans.org/">
-    <param name="text" value="<html><u>http://www.netbeans.org/</u></html>">
-    <param name="textFontSize" value="medium">
-    <param name="textColor" value="blue">
-</object>
-To create a link to a help set from another module, you need to know the code name base and path, e.g.:
-<a href="nbdocs://org.netbeans.modules.usersguide/org/netbeans/modules/usersguide/configure/configure_options.html">Using the Options Window</a>
-(This link will behave sanely if that module is disabled or missing.)
--->
-
