--- ide/editor.autosave/translatedfiles/javahelp/org/netbeans/modules/editor/autosave/docs/autosave-about_pt_BR.html
@@ -1,32 +1,0 @@
-<!--
-
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
-
--->
-<html>
-    <head>
-        <title>Sobre Salvar automaticamente</title>
-        <link rel="stylesheet" href="nbdocs:/org/netbeans/modules/usersguide/ide.css" type="text/css">
-        <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-    </head>
-    <body>
-        <h2>Sobre Salvar automaticamente</h2>
-         <p>Este módulo provê a funcionalidade de salvar automaticamente todos os arquivos não salvos abertos.</p>
-        <p>É possível ainda definir se este recurso estará habilitado ou não, bem como o intervalo de tempo entre as operações de salvar. Para alterar estas opções, favor utilizar as opções em <b>Ferramentas &gt; Opções &gt; Editor &gt; Salvar automaticamente</b>. Para habilitar/desabilitar este recurso, é possível utilizar-se ainda a opção <b>Arquivo &gt; <img src="auto_save.png"/> Salvar automaticamente</b>.</p>
-    </body>
-</html>
