+++ ide/editor.autosave/translatedfiles/src/org/netbeans/modules/editor/autosave/Bundle_pt_BR.properties
@@ -20,7 +20,7 @@ AdvancedOption_Tooltip=Configura\u00e7\u00f5es para o recurso de salvamento auto
 CTL_AutoSave=Salvar automaticamente
 OpenIDE-Module-Display-Category=Edi\u00e7\u00e3o
 OpenIDE-Module-Long-Description=Salva automaticamente todos os arquivos n\u00e3o salvos dentro de um intervalo de tempo predefinido.\nAs op\u00e7\u00f5es do m\u00f3dulo podem ser personalizadas atrav\u00e9s da se\u00e7\u00e3o 'Salvar automaticamente' em Ferramentas > Op\u00e7\u00f5es > Miscel\u00e2neas. Para habilitar/desabilitar este recurso, \u00e9 poss\u00edvel utilizar-se ainda a op\u00e7\u00e3o Arquivo > Salvar automaticamente.
-OpenIDE-Module-Name=Editor Save Automatically
+OpenIDE-Module-Name=Editor Autosave
 
 AutoSavePanel.jLabel1.text=Salvar arquivos a cada
 AutoSavePanel.jLabel2.text=minuto(s).
